requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
neo4j==5.14.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pydantic==2.5.0
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.24.3
networkx==3.2.1
tqdm==4.66.1
python-multipart==0.0.6